edit and delete trick
