+++ templates/trick/_form.html.twig
@@ -1,31 +1,50 @@
-{{ form_start(trick, {'method': 'POST'}) }}
+{{ form_start(trickForm) }}
 
-    {{ form_widget(trick._token) }}
+    {{ form_widget(trickForm._token) }}
     <div class="mb-3">
         <label for="name" class="form-label">Trick's name</label>
-        {{ form_widget(trick.name) }}
-        {{ form_errors(trick.name) }}
+        {{ form_widget(trickForm.name) }}
+        {{ form_errors(trickForm.name) }}
     </div>
     <div class="mb-3">
         <label for="description" class="form-label">Description</label>
-        {{ form_widget(trick.description) }}
-        {{ form_errors(trick.description) }}
+        {{ form_widget(trickForm.description) }}
+        {{ form_errors(trickForm.description) }}
     </div>
-    <div id="medias" class="mb-3" data-prototype="{{ include("trick/_media.html.twig", {form: trick.medias.vars.prototype})|e("html_attr") }}" data-index="{{ trick.medias|length > 0 ? trick.medias|last.vars.name + 1 : 0 }}">
-        {% for media in trick.medias %}
+    {% if app.request.attributes.get('_route') == 'app_trick_edit' %}
+		{% for media in trick.medias %}
+            {% if media.type is same as("Image") %}
+                <div id="mediasPreview" class="d-flex flex-column justify-content-center mx-2">
+                    <label for="mediaImage" class="mb-2">Media Image</label>
+                    <img src="{{ asset('/uploads/' ~ media.fileName)}}" alt="{{media.alt}}" width="150" class="rounded">
+                    <div>
+                        <a class="btn btn-sm btn-danger my-2" href="{{ path('app_trick_delete_media', {'id': media.id}) }}" data-delete="" data-token="{{ csrf_token('delete' ~ media.id) }}">Delete</a>
+                    </div>
+                </div>
+            {% else %}
+                <div id="mediasPreview" class="d-flex flex-column justify-content-center mx-2">
+                    <label for="mediaVideo" class="mb-2">Media Video</label>
+                    <p>{{media.url}}</p>
+                    <div>
+                        <a class="btn btn-sm btn-danger my-2" href="{{ path('app_trick_delete_media', {'id': media.id}) }}" data-delete="" data-token="{{ csrf_token('delete' ~ media.id) }}">Delete</a>
+                    </div>
+                </div>
+            {% endif %}
+		{% endfor %}
+	{% endif %}
+    <div id="medias" class="mb-3" data-prototype="{{ include("trick/_media.html.twig", {form: trickForm.medias.vars.prototype})|e("html_attr") }}" data-index="{{ trickForm.medias|length > 0 ? trickForm.medias|last.vars.name + 1 : 0 }}">
+        {% for media in trickForm.medias %}
             {{ include("trick/_media.html.twig", {form: media}) }}
         {% endfor %}
     </div>
     <button type="button" class="btn btn-primary btn-new mb-3" data-collection="#medias">New media</button>
-    {# {{ form_errors(trick.medias) }} #}
     <div class="mb-3">
         <label for="groups" class="form-label">Groups</label>
-        {{ form_widget(trick.groups) }}
-        {{ form_errors(trick.groups) }}
+        {{ form_widget(trickForm.groups) }}
+        {{ form_errors(trickForm.groups) }}
     </div>
-    {# {{ form_widget(trick.save) }} #}
     <button class="btn btn-primary mb-3">{{ button_label|default('Save') }}</button>
-{{ form_end(trick, { render_rest: false }) }}
+{{ form_end(trickForm, { render_rest: false }) }}
 
 {% block javascript %}
     <script src="{{ asset('assets/js/media.js') }}" type="text/javascript"></script>
