+++ templates/trick/edit.html.twig
@@ -11,6 +11,6 @@
 
 		{{ include('trick/_form.html.twig', {'button_label': 'Update'}) }}
 
-		<a href="{{ path('app_trick_index') }}" class="btn   btn-secondary">Back to list</a>
+		<a href="{{ path('app_home') }}" class="btn   btn-secondary">Back to list</a>
 	</div>
 {% endblock %}
